Moved auth files to Config.Auth project.
