+++ src/Config.Api/Program.cs
@@ -1,7 +1,7 @@
 global using pote.Config.Parser;
 global using pote.Config.Shared;
 using Df.ServiceControllerExtensions;
-using pote.Config.Api.Authentication;
+using pote.Config.Auth;
 using pote.Config.DataProvider.File;
 using pote.Config.DataProvider.Interfaces;
 
@@ -31,7 +31,7 @@ builder.Services.AddScoped<IDataProvider, DataProvider>();
 builder.Services.AddScoped<IApplicationDataAccess, ApplicationDataAccess>();
 builder.Services.AddScoped<IEnvironmentDataAccess, EnvironmentDataAccess>();
 
-builder.Services.AddTransient<IApiKeyValidation, ApiKeyValidation>();
+builder.Services.AddScoped<IApiKeyValidation, ApiKeyValidation>();
 builder.Services.AddScoped<ApiKeyAuthenticationFilter>();
 
 var app = builder.Build();
