+++ src/Config.Auth/ApiKeyAttribute.cs
@@ -1,0 +1,12 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace pote.Config.Auth;
+
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Method)]
+public class ApiKeyAttribute : ServiceFilterAttribute
+{
+    public ApiKeyAttribute()
+        : base(typeof(ApiKeyAuthenticationFilter))
+    {
+    }
+}
