+++ src/Config.sln
@@ -23,6 +23,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Config.DataProvider.Interfa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Config.Encryption", "Config.Encryption\Config.Encryption.csproj", "{10E10BC9-B4F8-4A20-AE59-A1AA3467DD38}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Config.Auth", "Config.Auth\Config.Auth.csproj", "{CE88A4F3-50A8-4E75-83D3-7D04CE43E5D5}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -69,6 +71,10 @@ Global
 		{10E10BC9-B4F8-4A20-AE59-A1AA3467DD38}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{10E10BC9-B4F8-4A20-AE59-A1AA3467DD38}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{10E10BC9-B4F8-4A20-AE59-A1AA3467DD38}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE88A4F3-50A8-4E75-83D3-7D04CE43E5D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE88A4F3-50A8-4E75-83D3-7D04CE43E5D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE88A4F3-50A8-4E75-83D3-7D04CE43E5D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE88A4F3-50A8-4E75-83D3-7D04CE43E5D5}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
